refactor: Clean up some imports and miscellaneous
