+++ global.json
@@ -1,7 +1,7 @@
 {
   "sdk": {
     "version": "8.0.201",
-    "rollForward": "latestPatch",
+    "rollForward": "latestFeature",
     "allowPrerelease": true
   }
 }
