+++ src/Api/Routes/Weather/Endpoints/DeleteWeather.cs
@@ -1,8 +1,6 @@
-using Api.Routes.Weather.Models;
-using Application.Weather.AddForecast;
+using Application.Weather.DeleteForecast;
 using MediatR;
 using Microsoft.AspNetCore.Http.HttpResults;
-using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Routes.Weather.Endpoints;
