+++ src/Application/Weather/DeleteForecast/DeleteForecastCommand.cs
@@ -1,6 +1,6 @@
 using FluentResults;
 using MediatR;
 
-namespace Application.Weather.AddForecast;
+namespace Application.Weather.DeleteForecast;
 
 public record DeleteForecastCommand(DateOnly Date) : IRequest<Result>;
